+++ src/quickPicks/quickPickSetupOpenai.ts
@@ -162,12 +162,14 @@ export async function quickPickSetupOpenai(
   ): Promise<string | undefined> {
     const OPENAI_APIKEY_MIN_LENGTH = 1
     const OPENAI_APIKEY_STARTSWITH = 'sk-'
-    const OPENAI_OAUTH2_STARTSWITH = 'ey'
+    const OPENAI_OAUTH2_BEARER_STARTSWITH = 'Bearer'
+    const OPENAI_OAUTH2_TOKEN_STARTSWITH = 'ey'
 
     // Native openai service key or oauth2 token
     return name.length >= OPENAI_APIKEY_MIN_LENGTH &&
       (name.startsWith(OPENAI_APIKEY_STARTSWITH) ||
-        name.startsWith(OPENAI_OAUTH2_STARTSWITH))
+        name.startsWith(OPENAI_OAUTH2_BEARER_STARTSWITH) ||
+        name.startsWith(OPENAI_OAUTH2_TOKEN_STARTSWITH))
       ? undefined
       : 'Invalid Api-Key or Token'
   }